requirements.txt
pillow==9.4.0
pytesseract==0.3.10
